more plugin parts; still need to implement actual plugin glue and integrate with DataSource
